requirements.txt
# Main dependencies
fastapi>=0.115.8,<0.116.0
uvicorn[standard]>=0.34.0,<0.35.0
python-multipart>=0.0.20,<0.1.0
ffmpeg-python>=0.2.0,<0.3.0
numpy<2.0.0
openai-whisper>=20240930
faster-whisper>=1.1.0,<2.0.0
whisperx>=3.3.1,<4.0.0
tqdm>=4.67.1,<5.0.0
llvmlite>=0.44.0,<0.45.0
numba>=0.61.0,<0.62.0
zhconv